+++ src/GRA.Data.SQLite/Migrations/SQLiteContextModelSnapshot.cs
@@ -13,7 +13,7 @@ namespace GRA.Data.SQLite.Migrations
         protected override void BuildModel(ModelBuilder modelBuilder)
         {
             modelBuilder
-                .HasAnnotation("ProductVersion", "1.1.5");
+                .HasAnnotation("ProductVersion", "1.1.6");
 
             modelBuilder.Entity("GRA.Data.Model.Answer", b =>
                 {
@@ -1458,8 +1458,6 @@ namespace GRA.Data.SQLite.Migrations
                         .HasMaxLength(150);
 
                     b.Property<int?>("BeforeRegistrationPage");
-
-                    b.Property<bool>("CollectPreregistrationEmails");
 
                     b.Property<DateTime>("CreatedAt");